requirements.txt
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
matplotlib==3.7.2
jupyter==1.0.0
requests==2.32.5
python-dotenv==1.2.1